Pequeñas mejoras y reorganización de carpetas
